Made couple of visual changes
I fixed the display of the inventory system so it deletes items and displays hilts properly.  I also changes the layout of the stations.  I also fixed and polished the visuals of the assembly minigame.  I also fixed the opening animation of the furnace minigame so it uses the raw items instead of the ingots.